feat: advances edit, perf optimisations, and misc fixes
- Advances: add PUT /api/advances/:id edit endpoint with amount/months update and monthly_amt recalculation; add updateAdvance API fn and edit modal in AdvancesPage
- Perf: replace ATS pipeline N+1 (60+ queries) with single GROUP BY FILTER query
- Perf: replace ATS report N+1 (7×N queries) with bulk candidate stage-count query
- Perf: rewrite buildPayrollRecords to pre-fetch attendance/advances/TDS in parallel before employee loop (200+ queries → 5+N)
- Perf: LMS reports skip CROSS JOIN when user_id/course_id filter is active
- Perf: add 6 DB indexes on hot columns (candidates, attendance, advances, payroll_records)
- Perf: increase Postgres pool max from 2 to 10
- Perf: add 45s TTL GET cache in apiFetch; busted on any mutation
- Perf: React.lazy route-level code splitting for all 22 pages
- Dashboard: surface API errors instead of silent empty state

+++ backend/src/routes/attendance/attendance.ts
@@ -59,7 +59,7 @@ router.get('/report', authenticateToken, requireRole('admin', 'hr'), async (req:
       COUNT(CASE WHEN a.status = 'leave'   THEN 1 END) AS leave_days,
       COUNT(CASE WHEN a.status = 'absent'  THEN 1 END) AS absent,
       COUNT(a.id) AS total_days,
-      ROUND(AVG(CASE WHEN a.work_hours IS NOT NULL THEN a.work_hours END), 1) AS avg_hours,
+      ROUND(AVG(CASE WHEN a.work_hours IS NOT NULL THEN a.work_hours END)::numeric, 1) AS avg_hours,
       (SELECT status FROM attendance WHERE user_id = u.id AND date = ?) AS today_status
     FROM users u
     LEFT JOIN attendance a ON u.id = a.user_id AND a.date LIKE ?
